requirements.txt
Flask==2.3.2
werkzeug==3.0.0
pymysql==1.0.2
requests==2.31.0
reportlab==4.0.4
beautifulsoup4==4.12.2
Flask-Mail==0.9.1
pyfiglet==0.8.post1
python-zap-v2.4==0.0.14
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2